+++ src/components/MetadataCollections/CollectionTechnical/CollectionTechnicalForm.js
@@ -1,11 +1,11 @@
 import PropTypes from 'prop-types';
 import { Field } from 'react-final-form';
-import { FieldArray } from 'react-final-form-arrays';
 import {
   FormattedMessage,
   injectIntl,
 } from 'react-intl';
 
+import { RepeatableTextField } from '@folio/stripes-leipzig-components';
 import {
   Accordion,
   Col,
@@ -15,7 +15,6 @@ import {
 } from '@folio/stripes/components';
 
 import BasicCss from '../../BasicStyle.css';
-import RepeatableField from '../../DisplayUtils/RepeatableField';
 import {
   Required,
   ValidateUrl,
@@ -67,10 +66,8 @@ const CollectionTechnicalForm = ({
         </Row>
         <Row>
           <Col xs={12}>
-            <FieldArray
-              component={RepeatableField}
+            <RepeatableTextField
               fieldValidate={ValidateUrl}
-              id="display_tickets"
               name="tickets"
               placeholder={intl.formatMessage({ id: 'ui-finc-config.collection.placeholder.tickets' })}
             />
@@ -86,10 +83,8 @@ const CollectionTechnicalForm = ({
         </Row>
         <Row>
           <Col xs={12}>
-            <FieldArray
-              component={RepeatableField}
+            <RepeatableTextField
               fieldValidate={ValidateUrl}
-              id="display_contentFiles"
               name="contentFiles"
               placeholder={intl.formatMessage({ id: 'ui-finc-config.collection.placeholder.contentFiles' })}
             />
@@ -105,9 +100,7 @@ const CollectionTechnicalForm = ({
         </Row>
         <Row>
           <Col xs={12}>
-            <FieldArray
-              component={RepeatableField}
-              id="display_solrMegaCollections"
+            <RepeatableTextField
               isFirstFieldRequired
               name="solrMegaCollections"
             />